+++ src/splitter/memory.rs
